Have ImageView resize to Image better
Make it so that the size of the imageView is determined by the image. This way it is not
necessary to add the frame to the image -- which wasn't ideal and also led to other
unresolved complications.

Now the imageView size matches the aspect ratio of the image, and so everything
lays out naturally.