Image border (#8)
* add imageView example

* add preliminary image

* add border only around image

* unify init some and show border on both object types

* correct init and reorganize code

* Have ImageView resize to Image better

Make it so that the size of the imageView is determined by the image. This way it is not
necessary to add the frame to the image -- which wasn't ideal and also led to other
unresolved complications.

Now the imageView size matches the aspect ratio of the image, and so everything
lays out naturally.

* update CollectionViewExample to use Anchorage completely

* update SingleCanvasViewExample to use Anchorage completely

* update MultipleCanvasViewExample to use Anchorage completely

* No longer add image to SelectionShowingView

This is no longer needed as the imageView matches the size of the image. The
image was being added to the SelectionShowingView because the approach
was going to require image manipulation that I didn't want to do on the
actual image. This work is no longer necessary.

* remove unnecessary translatesAutoresizingMask

* update CanvasController to use Anchorage completely

* update CenterIndicatingViews&UIViewExtension to use Anchorage completely

* add explanatory comment to example